refactor: 저번 내용 리팩토링 코드
