+++ src/pages/home/contexts/RecordingContext/RecordingModal/AudioVisualizer/AudioVisualizer.tsx
