+++ src/pages/home/contexts/RecordingContext/RecordingModal/AudioVisualizer/index.ts
