+++ src/utils/contexts/index.ts
@@ -1,8 +1,8 @@
 export * from './AlertContext';
 export * from './ApiContext';
 export * from './BottomNavContext';
-export * from './ResetContexts';
-export * from './SSEContexts';
-export * from './MemoContext';
-export * from './TagContext';
-export * from './RecordingContext';
+export * from '../../pages/home/contexts/ResetContexts';
+export * from '../../pages/home/contexts/SSEContexts';
+export * from '../../pages/home/contexts/MemoContext';
+export * from '../../pages/home/contexts/TagContext';
+export * from '../../pages/home/contexts/RecordingContext/RecordingContext';
